refactored

+++ 2-do_deploy_web_static.py
@@ -2,29 +2,56 @@
 """Generates a .tgz archive"""
 
 from fabric.api import put, run, env
+import os
 
 env.hosts = ['54.242.168.59', '52.206.252.73']
-env.user = "ubuntu"
+env.user = 'ubuntu'
 env.key_filename = "~/.ssh/id_rsa"
 
 
 def do_deploy(archive_path):
-    """ Deploy archive to server """
-    fd = archive_path.split("/")[1]
+    """Distributes an archive to web servers
+    Args:
+        archive_path: The path to the archive file
+
+    Returns:
+        True, Otherwise - False.
+    """
+
+    # Check if file at the path archive_path doesn’t exist
+    if not os.path.exists(archive_path):
+        return False
+
     try:
-        put(archive_path, "/tmp/{}".format(fd))
-        run("rm -r /data/web_static/releases/{}".format(fd))
-        run("mkdir -p /data/web_static/releases/{}".format(fd))
-        run("tar -xzf /tmp/{} -C /data/web_static/releases/{}".format(fd, fd))
-        run("rm /tmp/{}".format(fd))
-        run("mv /data/web_static/releases/{}/web_static/*\
-        /data/web_static/releases/{}/".format(fd, fd))
-        run("rm -rf /data/web_static/releases/{}/web_static".format(fd))
+        archive_file = archive_path.split("/")[-1]
+        arch_name = archive_file.split(".")[0]
+        releases_path = "/data/web_static/releases/{}".format(arch_name)
+        tmp_arch = "/tmp/{}".format(archive_file)
+
+        # Upload the archive to the /tmp/ directory of the web server
+        put(archive_path, "/tmp/")
+
+        # Create the release directory for ex: web_static_20170315003959
+        run("mkdir -p {}/".format(releases_path))
+
+        # Uncompress the archive
+        run("tar -xzf {} -C {}/".format(tmp_arch, releases_path))
+
+        # Delete the archive file
+        run("rm {}".format(tmp_arch))
+
+        # Move all the extracted content to our release directory
+        run("mv {}/web_static/* {}/".format(releases_path, releases_path))
+
+        # Remove the web_static directory
+        run("rm -rf {}/web_static".format(releases_path))
+
+        # Delete symbolic link
         run("rm -rf /data/web_static/current")
-        run("ln -s /data/web_static/releases/{}/\
-        /data/web_static/current".format(fd))
-        print("New version deployed!")
+
+        # Create a new the symbolic link
+        run("ln -s {}/ /data/web_static/current".format(releases_path))
+
         return True
     except Exception:
-        print("Deployment failed!")
         return False
